+++ firmware/components/midi_mapping/include/midi_mapping_json_builder.h
@@ -19,41 +19,17 @@ namespace midi {
 
 // TODO Document should be const, but we can't get it's allocator when it is.
 // Maybe just document that implementations must not change the document?
+// Maybe pass the allocator instead of the document?
 template<typename T>
 rapidjson::Value to_json(rapidjson::Document &document, const T &object);
+
+template<>
+rapidjson::Value to_json(rapidjson::Document &document, const Mapping &object);
 
 #if 0
 template<>
 rapidjson::Value to_json(rapidjson::Document &document, const GetResponse &object)
-{
-    rapidjson::Value json;
-    json.SetObject();
-
-    return json;
-}
 #endif
-
-// TODO put full specialisations in a cpp file to avoid ODR violation
-template<>
-rapidjson::Value to_json(rapidjson::Document &document, const Mapping &object)
-{
-    rapidjson::Value json;
-    json.SetObject();
-
-    rapidjson::Value midi_channel;
-    midi_channel = object.midi_channel;
-    json.AddMember("midi_channel", midi_channel, document.GetAllocator());
-
-    rapidjson::Value cc_number;
-    cc_number = object.cc_number;
-    json.AddMember("cc_number", cc_number, document.GetAllocator());
-
-    rapidjson::Value parameter_id;
-    parameter_id.SetString(object.parameter_name.data(), object.parameter_name.size(), document.GetAllocator());
-    json.AddMember("parameter_id", parameter_id, document.GetAllocator());
-
-    return json;
-}
 
 }
 }
